refactor articlePopover
