+++ openaev-front/src/admin/components/common/articles/ArticleForm.tsx
@@ -293,7 +293,7 @@ const ArticleForm = ({
         )}
       />
 
-      <Grid container spacing={3} style={{ marginTop: 0 }}>
+      <Grid container spacing={3} style={{ marginTop: 20 }}>
         <Grid size={{ xs: 4 }}>
           <TextFieldController
             name="article_comments"
